requirements.txt
mistune
wheel>=0.24.0
Django==1.11.27
xxhash>=0.6.1
celery[redis]
django-crispy-forms>=1.6.0
six
pillow
djangorestframework
PyMySQL
mysqlclient
django-tinymce
bs4
django-simpleui==3.6